requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.2
numpy>=1.23.0
datasets>=2.12.0
tqdm>=4.65.0
torchviz>=0.0.2
pyyaml>=6.0